Factor out eth bridge router from the shell router
